Refactor scrapers into modules
